refactor(root): changed retrieval method of current user
The previous way of retrieving was not optimal as it dynamically
rerenders all the time. And so a new function was made,
getCurrentUser().

This new function allows to retrieve the user in a cached manner and
only rendering if data is changed.
