+++ app/(root)/account/components/forms/AccountSettingPasswordForm.tsx
@@ -1,11 +1,11 @@
 "use client";
 
+import { changePassword } from "@/auth/actions";
 import CustomFormField, {
   FormFieldType,
   InputType,
 } from "@/components/shared/CustomFormField";
 import { Form } from "@/components/ui/form";
-import { changePassword } from "@/lib/actions/auth.actions";
 import { formatErrorResponse } from "@/lib/formatters";
 import { cn } from "@/lib/utils";
 import { ChangePasswordFormData, changePasswordFormSchema } from "@/schemas";
