+++ app/(root)/distributions/page.tsx
@@ -2,20 +2,20 @@ import DistributionClient from "./DistributionClient";
 import { getBrands } from "@/lib/actions/product.classications.actions";
 import { getProducts } from "@/lib/actions/product.actions";
 import { getDistributions } from "@/lib/actions/distribution.actions";
-import { getSession } from "@/lib/session";
 import { getAssets } from "@/lib/actions/asset.actions";
+import { getCurrentUser } from "@/auth/currentUser";
 
 export default async function Distributions() {
   const brands = await getBrands();
   const products = await getProducts();
   const distributions = await getDistributions();
   const assets = await getAssets();
-  const userSession = await getSession();
+  const employee = await getCurrentUser({ withEmployeeProfile: true });
 
   return (
     <DistributionClient
-      user={userSession!.user}
-      employee={userSession?.employee?.name || "Admin"}
+      user={employee.user}
+      employee={employee.first_name || "Admin"}
       distributions={distributions}
       brands={brands}
       products={products}
