+++ app/(root)/page.tsx
@@ -5,15 +5,14 @@ import {
   getProductsAboutToExpireCount,
   getProductsExpiredCount,
   getRemainingTaskCount,
-  getTotalItemStock,
 } from "@/lib/actions/metrics.actions";
-import { getSession } from "@/lib/session";
 import { InsightsMetrics } from "@/types/metrics";
 import HomeClient from "./Client";
+import { getCurrentUser } from "@/auth/currentUser";
 
 export default async function Home() {
   const distributions = await getDistributions();
-  const user = (await getSession())!.user;
+  const user = await getCurrentUser();
   const metrics: InsightsMetrics = {
     monthlyDistributionFlow: await getMonthlyDistributionFlow(),
     remainingTaskCount: await getRemainingTaskCount(),
@@ -24,7 +23,7 @@ export default async function Home() {
 
   return (
     <HomeClient
-      user={user}
+      user={user!}
       distributions={distributions}
       metrics={metrics}
       distributionFlowComparison={distributionFlowComparison}
